+++ packages/eyes-sdk-core/test/utils/MockDriver.js
@@ -9,6 +9,11 @@ const WELL_KNOWN_SCRIPTS = {
   'dom-capture': script => /^\/\* @applitools\/dom-capture@[\d.]+ \*\//.test(script),
 }
 
+const DEFAULT_DESKTOP_UA =
+  'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.89 Safari/537.36'
+const DEFAULT_MOBILE_UA =
+  'Mozilla/5.0 (Linux; Android 8.0.0; Pixel 2 XL Build/OPD1.170816.004) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.135 Mobile Safari/537.36'
+
 const DEFAULT_STYLES = {
   'border-left-width': '0px',
   'border-top-width': '0px',
@@ -22,9 +27,11 @@ const DEFAULT_PROPS = {
 }
 
 class MockDriver {
-  constructor({isNative = false, isMobile = false} = {}) {
-    this._isNative = isNative
-    this._isMobile = isMobile
+  constructor({device, platform, browser, ua} = {}) {
+    this._device = device
+    this._platform = platform
+    this._browser = browser
+    this._ua = ua
     this._window = {
       title: 'Default Page Title',
       url: 'http://default.url',
@@ -113,7 +120,8 @@ class MockDriver {
       return 1
     })
     this.mockScript(snippets.getUserAgent, () => {
-      return 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.89 Safari/537.36'
+      if (this._ua) return this._ua
+      return this.info.isMobile ? DEFAULT_MOBILE_UA : DEFAULT_DESKTOP_UA
     })
     this.mockScript(snippets.getViewportSize, () => {
       return {width: this._window.rect.width, height: this._window.rect.height}
@@ -202,12 +210,24 @@ class MockDriver {
       }
     }
   }
+
+  get info() {
+    return {
+      isMobile: this._device ? Boolean(this._device.isMobile) : false,
+      isNative: this._device ? Boolean(this._device.isNative) : false,
+      deviceName: this._device ? this._device.name : null,
+      platformName: this._platform ? this._platform.name : null,
+      platformVersion: this._platform ? this._platform.version : null,
+      browserName: this._browser ? this._browser.name : null,
+      browserVersion: this._browser ? this._browser.version : null,
+    }
+  }
   async executeScript(script, args = []) {
+    if (this.info.isNative) throw new Error("Native context doesn't support this method")
     args = serialize(args)
     let result = this._scripts.get(script)
     if (!result) {
       const name = Object.keys(WELL_KNOWN_SCRIPTS).find(name => WELL_KNOWN_SCRIPTS[name](script))
-      console.log
       if (!name) return null
       result = this._scripts.get(name)
     }
@@ -255,15 +275,15 @@ class MockDriver {
     Object.assign(this._window.rect, rect)
   }
   async getUrl() {
-    if (this._isNative) throw new Error("Native context doesn't support this method")
+    if (this.info.isNative) throw new Error("Native context doesn't support this method")
     return this._window.url
   }
   async getTitle() {
-    if (this._isNative) throw new Error("Native context doesn't support this method")
+    if (this.info.isNative) throw new Error("Native context doesn't support this method")
     return this._window.title
   }
   async visit(url) {
-    if (this._isNative) throw new Error("Native context doesn't support this method")
+    if (this.info.isNative) throw new Error("Native context doesn't support this method")
     this._window.url = url
   }
   async takeScreenshot() {