+++ packages/sdk-shared/coverage-tests/custom/TestDisableBrowserFetching.spec.js
@@ -10,7 +10,7 @@ const adjustUrlToDocker = require('../util/adjust-url-to-docker')
 
 describe('TestDisableBrowserFetching', () => {
   let testServer
-  let driver
+  let driver, destroyDriver
 
   before(async () => {
     const staticPath = path.join(__dirname, '../fixtures')
@@ -26,11 +26,11 @@ describe('TestDisableBrowserFetching', () => {
   })
 
   beforeEach(async () => {
-    driver = await spec.build({browser: 'chrome'})
+    ;[driver, destroyDriver] = await spec.build({browser: 'chrome'})
   })
 
   afterEach(async () => {
-    await spec.cleanup(driver)
+    await destroyDriver()
   })
 
   it('sends dontFetchResources to dom snapshot', async () => {