+++ packages/sdk-shared/coverage-tests/custom/TestEyesDifferentRunners/TestEyesDifferentRunnersVG_part2.spec.js
@@ -8,12 +8,12 @@ const {testSetup, getCheckSettings, validateVG} = require('./EyesDifferentRunner
 
 describe('TestEyesDifferentRunners VG', () => {
   afterEach(async function() {
-    await spec.cleanup(this.webDriver)
+    await this.destroyDriver()
     await this.eyes.abortIfNotClosed()
   })
 
   beforeEach(async function() {
-    this.webDriver = await spec.build({browser: 'chrome'})
+    ;[this.webDriver, this.destroyDriver] = await spec.build({browser: 'chrome'})
     this.eyes = await getEyes({isVisualGrid: true})
     let conf = this.eyes.getConfiguration()
     conf.setTestName(`Top Sites - ${this.currentTest.title}`)