+++ packages/sdk-shared/coverage-tests/generic/supported-tests.js
@@ -331,4 +331,5 @@ module.exports = [
   {name: 'CheckWindowOnJsLayoutPage', executionMode: {isVisualGrid: true}},
   {name: 'CheckWindowOnJsLayoutPageWithConfig', executionMode: {isVisualGrid: true}},
   {name: 'CheckRegionWithFractionalMetrics', executionMode: {isCssStitching: true}},
+  {name: 'CheckHoveredRegionInViewport', executionMode: {isScrollStitching: true}},
 ]