+++ packages/eyes-playwright/CHANGELOG.md
@@ -10,5 +10,5 @@
 - updated to @applitools/visual-grid-client@14.6.4 (from 14.6.3)
 - updated to @applitools/eyes-sdk-core@12.2.0 (from 12.0.1)
 - updated to @applitools/visual-grid-client@15.0.0 (from 14.6.4)
-- updated to @applitools/eyes-sdk-core@12.2.3 (from 12.2.0)
-- updated to @applitools/visual-grid-client@15.0.3 (from 15.0.0)
+- updated to @applitools/eyes-sdk-core@12.2.4 (from 12.2.0)
+- updated to @applitools/visual-grid-client@15.0.4 (from 15.0.0)