+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## Unreleased
+- updated to @applitools/eyes-sdk-core@12.2.4 (from 12.2.0)
+- updated to @applitools/visual-grid-client@15.0.4 (from 15.0.0)
 
 
 ## 2.22.0 - 2020/8/30